Text.JSON.Canonical

Description

An implementation of Canonical JSON.

http://wiki.laptop.org/go/Canonical_JSON

The "canonical JSON" format is designed to provide repeatable hashes of JSON-encoded data. It is designed for applications that need to hash, sign or authenitcate JSON data structures.

The format is an extended subset of the normal JSON format.

Canonical JSON is parsable with any full JSON parser, and it allows whitespace for pretty-printed human readable presentation, but it can be put into a canonical form which then has a stable serialised representation and thus a stable hash.

The basic concept is that a file in the canonical JSON format can be read using parseCanonicalJSON. Note that this input file does not itself need to be in canonical form, it just needs to be in the canonical JSON format. Then the renderCanonicalJSON function is used to render into the canonical form. This is then the form that can be hashed or signed etc.

The prettyCanonicalJSON is for convenience to render in a human readable style, since the canoncal form eliminates unnecessary white space which makes the output hard to read. This style is again suitable to read using parseCanonicalJSON. So this is suitable to use for producing output that has to be later hashed or otherwise checked.

Known bugs/limitations:

Decoding/encoding Unicode code-points beyond U+00ff is currently broken

Types

data JSValue

Constructors

JSNull

JSBool !Bool

JSNum !Int54

JSString !JSString

JSArray [JSValue]

JSObject [(JSString, JSValue)]

data Int54

54-bit integer values

JavaScript can only safely represent numbers between -(2^53 - 1) and 2^53 - 1.

TODO: Although we introduce the type here, we don't actually do any bounds checking and just inherit all type class instance from Int64. We should probably define fromInteger to do bounds checking, give different instances for type classes such as Bounded and FiniteBits, etc.

data JSString

Canonical JSON strings are in fact just bytes.

Parsing and printing

parseCanonicalJSON :: ByteString -> Either String JSValue

Parse a canonical JSON format string as a JSON value. The input string does not have to be in canonical form, just in the "canonical JSON" format.

Use renderCanonicalJSON to convert into canonical form.

renderCanonicalJSON :: JSValue -> ByteString

Render a JSON value in canonical form. This rendered form is canonical and so allows repeatable hashes.

For pretty printing, see prettyCanonicalJSON.

NB: Canonical JSON's string escaping rules deviate from RFC 7159 JSON which requires

"All Unicode characters may be placed within the quotation marks, except for the characters that must be escaped: quotation mark, reverse solidus, and the control characters (U+0000 through U+001F)."

Whereas the current specification of Canonical JSON explicitly requires to violate this by only escaping the quotation mark and the reverse solidus. This, however, contradicts Canonical JSON's statement that "Canonical JSON is parsable with any full JSON parser"

Consequently, Canonical JSON is not a proper subset of RFC 7159.

prettyCanonicalJSON :: JSValue -> String

Render a JSON value in a reasonable human-readable form. This rendered form is not the canonical form used for repeatable hashes, use renderCanonicalJSON for that.

Type classes

class ToJSON m a where

Methods

toJSON :: a -> m JSValue

class FromJSON m a where

Methods

fromJSON :: JSValue -> m a

class ToObjectKey m a where

Used in the ToJSON instance for Map

Methods

toObjectKey :: a -> m JSString

class FromObjectKey m a where

Used in the FromJSON instance for Map

Methods

fromObjectKey :: JSString -> m (Maybe a)

class (Applicative m, Monad m) => ReportSchemaErrors m where

Monads in which we can report schema errors

Methods

expected :: Expected -> Maybe Got -> m a

type Expected = String

type Got = String

expectedButGotValue :: ReportSchemaErrors m => Expected -> JSValue -> m a

Utility

toJSString :: String -> JSString

fromJSString :: JSString -> String

fromJSObject :: ReportSchemaErrors m => JSValue -> m [(JSString, JSValue)]

fromJSField :: (ReportSchemaErrors m, FromJSON m a) => JSValue -> JSString -> m a

Extract a field from a JSON object

fromJSOptField :: (ReportSchemaErrors m, FromJSON m a) => JSValue -> JSString -> m (Maybe a)

mkObject :: forall m. Monad m => [(JSString, m JSValue)] -> m JSValue