Text.JSON.Canonical.Types

Description

Documentation

data JSValue

Constructors

JSNull

JSBool !Bool

JSNum !Int54

JSString !JSString

JSArray [JSValue]

JSObject [(JSString, JSValue)]

newtype Int54

54-bit integer values

JavaScript can only safely represent numbers between -(2^53 - 1) and 2^53 - 1.

TODO: Although we introduce the type here, we don't actually do any bounds checking and just inherit all type class instance from Int64. We should probably define fromInteger to do bounds checking, give different instances for type classes such as Bounded and FiniteBits, etc.

Constructors

Int54

Fields

int54ToInt64 :: Int64

data JSString

Canonical JSON strings are in fact just bytes.

toJSString :: String -> JSString

fromJSString :: JSString -> String